# chore(queue): replace channel queue with ring buffer queue.